requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
python-dotenv>=1.0.0
requests>=2.31.0
websockets>=11.0.3
oandapyV20>=0.7.2
python-dateutil>=2.8.2 